WIP regfactoring
